requirements.txt
streamlit
groq
kagglehub
transformers
sentence-transformers
faiss-cpu
datasets